:safety_vest: Validating baseline by overfitting env
I'm running a experiment on the DQN baseline to check if the implementation is correct by only overfitting one environment. If it increases the reward, decreases the norm of gradients, and reduces loss, then the algorithm is correct and I can proceed with diversity in environments as well as implementing my curriculum algorithm.